// src/features/celeb-dash/engine/types.ts
export type Upgrades = Record<UpgradeId, number>;

export type UpgradeId = 'sneakers' | 'tote' | 'dj' | 'chef' | 'bodyguard' | 'glamsquad' | 'powerbank' | 'mesh';